Update gameClient.js to fix the Cookie Storm problem.
There is a classic async problem here. The setTimeout just tries to repop the golden cookie in 500 ms, but it doesn't take into account that the cookie might have already been popped within that time (by the autoclicker), which ends up meaning that you try to pop the cookie twice (which has a lot of bad side effects).

I fixed this by just making sure that the cookie hasn't been popped yet before popping it in the async function.